requirements.txt
pandas
numpy
plotly
streamlit
yfinance
textblob
schedule
tweepy
kafka-python
pyspark